+++ templates/data/loot-salvage.php
@@ -1,0 +1,1 @@
+<header class="entry-header"><h2 class="entry-title"><?php echo $title; ?><?php if(isset($count)) {echo ' (' . $count . ')';} ?></h2></header>