+++ templates/single-intel.php
@@ -19,6 +19,8 @@ switch($termObject['0']->slug) {
 		$dscanDataOnGrid = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-onGrid', true));
 		$dscanDataOffGrid = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-offGrid', true));
 		$dscanDataShipTypes = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-shipTypes', true));
+		$dscanUpwellStructures = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-upwellStructures', true));
+		$dscanDeployables = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-deployables', true));
 		$dscanDataSystem = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-system', true));
 		$dscanDataTime = \maybe_unserialize(\get_post_meta(\get_the_ID(), 'eve-intel-tool_dscan-time', true));
 
@@ -28,6 +30,8 @@ switch($termObject['0']->slug) {
 			'dscanDataOffGrid' => $dscanDataOffGrid,
 			'dscanDataShipTypes' => $dscanDataShipTypes,
 			'dscanDataSystem' => $dscanDataSystem,
+			'dscanUpwellStructures' => $dscanUpwellStructures,
+			'dscanDeployables' => $dscanDeployables,
 			'eveTime' => $dscanDataTime,
 		];